+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/queue/ProcessQueueWebSocketHandler.java
@@ -22,7 +22,6 @@ package com.walmartlabs.concord.server.process.queue;
 
 import com.walmartlabs.concord.server.PeriodicTask;
 import com.walmartlabs.concord.server.org.OrganizationDao;
-import com.walmartlabs.concord.server.org.OrganizationEntry;
 import com.walmartlabs.concord.server.org.project.RepositoryDao;
 import com.walmartlabs.concord.server.org.project.RepositoryEntry;
 import com.walmartlabs.concord.server.process.ProcessManager;